Move HTMLMap class to a diferent file
