+++ cartoframes/viz/html/__init__.py
@@ -1,0 +1,7 @@
+from __future__ import absolute_import
+
+from .HTMLMap import HTMLMap
+
+__all__ = [
+  'HTMLMap'
+]
